General revision of mainly evaluator.w
The evaluator now has 2 engines: the old one that will type-check and then
evaluate the expression that was generated by the parser, and a new one in
which type-checking produces new C++ objects that can execute themselves.

The old engine operates when verbose is off, the new one when verbose is on;
thus results can be compaared. This commit serves to fix the old engine in the
CVS repository in its final state, before it gets retired permanently.

Array subscription is now possible, for rows, vectors, and matrices; the
fullest functionality is availble using the new engine; that engine now also
has conversions between array types in all settings where it can be done.

A general clean-up has been performed, preparing for the great leap forward.

Finally copyright and GPL notices have finally been added to all source files.