Resolved problem of halt for input files without final newline
After input of a file not ending with a newline the program would halt. This
was due to faulty logic in |InputBuffer::getline|, which did not even enter
into the code that pops input files when |stream->good()| returned |false|
right away, as happens when getting the line _after_ the final line without
newline. The logic has  been thoroughly altered, and now handles even strange
scenarios as files ending with a backslash character gracefully. Another
change is that a failing nested include will now abandon all includes.