Modified |BufferedInput::push_file| to check the actual file name
The mechanism that avoids multiple inclusion of realex source files (when
using '<' rather thant '<<') used to check only the name specified by the
user, possibly without extension, causing for instance 'basic' and 'basic.rx'
to be considered two different files. Now only the name under which a file is
successfully opened is stored and tested (note that this does contain a
possible path to the file, so a file can still be multiply loaded by
specifying different paths to it). This would imply that even files that will
be skipped to avoid multiple inclusions are opened each time before doing the
test, only to be closed immediately afterwards; this behaviour is avoided if
the file name requested exactly matches one successfuly opened before (it will
be skipped without even looking if it still exists). This is an incentive to
specify the '.rx' extension explicitly when including from files. Files that
failed to open are not recorded as read; if the user creates the file during
the session of realex, it will be opened on the next attempt.