Added implicit conversions vec->ratvec, [int]->ratvec and polished scripts
The new automatic promotion of vectors and integer lists to ratvec made
several function overloads unnecessary, sometimes conficting, so they were
removed. One fundamental change to the scripts is to rename evaluation of
polynomials from the '%' operator to the function 'eval'; this avoids a
conflict with %^(vec,int) with meaning "reduce all entries modulo number".