Modify low-level input handling in strange cases, and give better feedback
Strange cases include situations where an included file ends in the middle of
a command. This is now signalled by the input buffer transmitting a form-feed
character to the lexer, which will not ignore it and break off any compound
construction like a comment that it is processing. Eventually the parser will
report the form-feed character as erroneous if not at the start of a command.

The switching of input files is also better tracable in the output written
during the include, by indenting it and reporting begin/end of includes. The
current solution may not yet be ideal because it gives unsuppressable output.