Changed atlas function |main| to do some command line argument processing
One can now pass a path to the message directory to override the one that was
fixed at compilation time, which makes it possible to install the files of the
atlas tree at a different location than where they were originally compiled.
Also a --version option is implemented to just find out the atlas version.