Add 'distribution' target to the makefile; also 'make all' builds atlas+realex
In fact the distribution will be a simplified version of the atlas sources,
with preprocessing by cwebx and bison already performed, so that our clients
need not have those programs installed. Also the 'make install' command in the
distribution will put the bare essentials (executables, messages, scripts and
stuff like LICENCE) in the installation directory, while putting symbolic
links or a shell script (to accommodate a moved message directory) in the
'binary' directory.