Some finalisation to the recent changes in the Makefile organisation.
Removed (almost) all references to atlas.exe in these files. Adapted the main
makefile to include the changes made in distr_Makefile. Notably, the atlas
executable, which has lost its .exe suffix, and the realex executable are both
placed at the root of the atlas directory, and of the install directory after
make install; the bin directory will get links to the install directory. The
automatic generation of dependencies for .o files from the interpreter
subdirectory is currently not done, as this caused a deadlock of the make
process: the corresponding .cpp files may not exist, and the makefile needs
_all_ dependencies before it even starts figuring how it could make those .cpp
files, so it cannot proceed. A very curious (unrelated) bug in the current
version is that 'make realex' _always_ re-links the program, even if it is
already up-to-date. Calling 'make --debug=basic realex' reports that the file
realex does not exist, in spite of the evidence to the contrary.

 LocalWords:  exe distr cpp pr rec jeff