Change 'stadardize' to noun 'standardization' when appropriate; simplify a bit
Some insignificant changes related to permutations, sorting, standardization.
When standardization in the sense of a particular permutation is meant, the
noun should be used. A comment describing it was corrected. Also the
functionality of exporting the 'stops' from standardization is now actually
used, avoiding some duplicated work in |Block_base::renumber_x|.