Extended syntax 'let .. , f(..) = ..' and 'let + = ..'; overload based on type
There was an incosistency in the syntax, allowing functions defined by the
usual 'f (...) =' syntax immeditely after 'let', but not after a comma; this
is now allowed as well. Also the behaviour of 'set' with function values has
changed: whereas only the special function syntax used to result in defining
an overload instance, it is now done whenever the value type allows it, namely
a function type with at least one argument. This syntax is now also allowed
with operators, for instance 'let + = ...' which facilitates defining
operators using a recusive function body; a separate identifier in the
recursion is still needed (one cannot an operator), but at least that
identifier can be defined locally (see the change to ^@(mat,int) in basic.rx).