New KL algorithm, avoiding Thickets
New methods newRecursion, newRecursionRow, lastklPol, firstNiceAscent,
firstImaginaryReal in kl::Helper class; and safeDivide, safeQuotient in
Safe_Poly class.

The method "safeQuotient" is sometimes passed a degree d which is
morally a half integer, but has been truncated to the next
lower integer. In this case the "assert base::degree() < d+1" is still
wrong; equality is allowed, in which case the added mu is supposed to
be zero, so that the result in the end does indeed have degree at most
d+1. This correction weakens the assert enough that everything ought
to work, but it still isn't testing what's supposed to happen as
carefully as it ought to.