Make 'string' subscriptable and comparable type; decomposition '%' of LieType
Added possibilities for a string s: subscription s[i] (returns a string of
length 1), looping over characters: for c in s do ... od, conversion of intial
character to ASCII code, and conversely conversion of code to a string of
length 1. Also defined comparison <= of strings (other relations in basic.rx).
For LieType no subscripting or looping is implemented (as this would require
evaluator.w to know about built-in types), but an operator '%' is added the
expands a lie type to row of Lie types, either simple or type T1. The function
nr_factors@LieType has been renamed '#', for consistency with similar cases.