Improve method lastklPol, simplify new algorithm by computing for all primitive
Firstly, |lastklPol| was improved by passing an index telling from which point
up there are actually elements to search for. But it was also ranged out.

Next the new algorithm was fundamentally simplified by computing for all
primitive x, not just extremal ones. Computing KL polynomials for primitive
non-extremal elements was done on the fly in the (now unused) lastklPol
method, but it is more efficient to compute them just once in
|newRecursionRow|. To tell the extremal and merely primitive cases apart, a
new method |ascent_descent| was added to KLsupport, which turns out to double
the |firstAscent| function, but more efficiently. Also the |safeQuotient|
method was simplified, but exploiting the fact that coefficients are
non-negative; passing the degree of the possibly missing term is now only used
in |assert| statements. Finally |Helper::muNewFormula| was simplified; it
turned to have trouble with being passed a list of extremal rather than
primitive elements, and the code to compensate for this was simply taken out.