Simplify the new algorithm by computing for all primitive x, not just extremal
Computing KL polynomials for primitive non-extremal elements was done on the
fly in the (now unused) lastklPol method, but it is more efficient to compute
them just once in |newRecursionRow|. To tell the extremal and merely primitive
cases apart, a new method |ascent_descent| was added to KLsupport, which turns
out to double the |firstAscent| function, but more efficiently. Also the
|safeQuotient| method was simplified, but exploiting the fact that
coefficients are non-negative; passing the degree of the possibly missing term
is now only used in |assert| statements. Finally |Helper::muNewFormula| was
simplified; it turned to have trouble with being passed a list of extremal
rather than primitive elements, and the code to compensate for this was simply
taken out (but it would have been easier to find out this aspect of its
functioning if the comments had been a bit more explicit).