Move mu-generation into |complete_primitives|, |fillMuRow| checks it for now
The goal is to make |fillMuRow| obsolete by doing the work at the same time as
storing the KL polynomials for extremal elements. For now this is only done
for those |y| with a direct recursion, which pass their KL polynomials
through |complete_primitives|, but the case of |newRecursionRow| is actually
more interesting as it needs those mu values right away. The case of length
one less than |y| is handled somewhat awkwardly, as the down-set of y needs to
be added (with coefficients 1), but it consists of elements that are not
extremal and in part even not primitive. Adding them to the end of the MuRow
may make it unsorted by block element; this is probably harmless because the
sortedness is not used, but for now the elements are insertion-sorted into the
list of extremal x's with nonzero mu, which is never a lot of work. Since this
is rather fragile stuff, we double-check for now in |fillMuRow| to get the
same result as the one we have always been using.