Final clean-up of kl.cpp: don't store the P_{y,y} that aren't ever looked at
Since the (indices of) polynomials $P_{y,y}=1$ were stored at the end of their
row, just after the polynomials for primitive $x$ with $l(x)<l(y)$, which is
not the location |self_index(y)| where $y$ appears in the list of primitive
elements for its own descent set, the method |klPol| had to (and still has to)
specially test for elements primitivizing to $y$ itself, But the final entries
of each |d_kl[y]| were therefore never even looked at, so they were removed
from the storage altogether. This did require a change to the filekl module,
which in fact was already put off-kilter by the recent change of storing
indices for zero polynomials explicitly. This code was adapted and testd.

A method |KLContext::KL_pol_index| was added for the convenience of realex
(function |KL_block|), which used to have a bug due to the fact that |klPol|
could return a |KLPolRef| to a copy of |One| with address not in |d_store|.

Also the old |makeExtremalRow| has become a value-returning function
|extremalRow|, and similarly |primitiveRow|; moreover these functions no
longer produce the value |y| at the end of the list (so their result only
depends on the length and the descent set of |y|); clients have been adapted.

Finally the class constants |KLContext::d_zero| and |KLContext::d_one| were
made into |enum| values, so they don't take up any actual storage.