Repair lingering bug in nonintegral block construction (upward Cayley)
The method |nblock_help::parent_up_Cayley| supposed that the |TorusElement|
values used to represent a |y| part is always integral on the roots of the
integral subsystem. However since it may itself modify that |TorusElement| by
half of a root, that assumption is invalid. So it now corrects by a rational
multiple of a root. Added testing of integrality on subsystm real roots.