Removed two bugs in the handling of tables to store representation results.
The first one involved reusing the KL tables of (older and maybe larger) block
in which incomparable elements may have switched order; this may lead to some
entries to be looked up on the wrong side of the diagaonal of the KL matrix,
giving out-of-bound access and therefore random values (instead of $0$). An
added test for out-of-bound access, replacint it with $0$, solves the problem.

The other bug was another case of a forgotten implicit |make_dominant|. This
happens in |Rep_table::deformation| which tries to store its result as
associated to its original paremeter snapped to its first reducibility point,
where |Rep_table::deformation_terms| should create an entry for it in the hash
table. However, the element looked up in |Rep_table::deformation_terms| is
really the top element of the block generated for the reducibility point,
which has an implicit |make_dominant| that has been applied before the block
was actually generated. Therefore one should in |Rep_table::deformation| also
apply |make_dominant| to the snapped parameter before hashing it for storage.