Major reorganisation of the blocks module, needed to define the block twist
Many changes were necessary to make computing the block involution given by
conjugation by delta work correctly for the various block constructors. This
had to deal with having a way to look up block elements by their value in the
various cases (including partial_block which does not have complete R-packets,
or collect such packets together, so they are of no use for look-up) and the
complications caused by renumbering of x values between parent and child. In
the end, the |parent_x| map was made the identity for partial blocks, so no
renumbering is needed, and hashing is used for look-up in non-integral blocks.

In passing, data structures in |Block_base| were changed to vector(s) of
structures, like KGB, and  the realex function |dual_KL| was revives,
documented, and made consistent in type with |raw_KL|.