Restructured the command handling
Every module like mainmode now defines a function producing a |CommandNode|
that is called during static initialisation from main.cpp to generate, in the
proper order, the |CommandTree| values, which establishes descendance
relations and ensures inheritance of commands from all ancestors. The commands
defined in the |CommandNode| itself cannot be overridden by a command of the
same name in a ancestor. The odd one out among the modes is the |helpNode|, as
its construction produces a side effect in a static variable |tagDict|, whence
it must be called from the same module to avoid initialisation fiasco;
fortunately this node does not have any ancestors so this works correctly.