Rework and simplify command loop thoroughly. Bump version number to 0.5.4
The module commands.cpp was overhauled, a more explicit search for a command
name is done, and entry and exit from modes is done in a less ad-hoc way.
Commands from sibling modes will now be found (unless they are masked by
commands in the current mode), dropping to the common parent mode first.