Make extblock safer with respect to partial and non-integral blocks
Block links that mathematically exist may nevertheless be UndefBlock in a
partial block; this change tries to make the extblock construction succeed
for such cases, even though maybe not all ascent statuses can be reliably
determined. For non-integral blocks a bug is fixed that took the folded
diagram to be obtained from the full root datum with involution, rather than
for the subdatum, while block links refer to simple roots of the subdatum.