Finally get the twist on the dual KGB (for blocktwist) right in all cases.
* files weyl.{h,cpp}

Introduce a |Twist| (diagram automorphism) field |Chevalley| in the class
|WeylGroup|, and method |Chevalley_dual| to access it. No root datum is at
hand during the construction of |WeylGroup|, so use conjugation by $w_0$.

Add methods |dual_twisted| (for generators and Weyl words) to the class
|TwistedWeylGroup| composes the stored twist with the Chevalley dual. The
method |dual_twist| that returns this composed automorphism as a |Twist|
value is simplified to use |WeylGroup::Chevalley_dual| (whereas it used to use
conjugation by $w_0$ is is now done during construction of the |WeylGroup|).

* files tits.{h,cpp}

Add field |TitsGroup::dual_involution| (a binary matrix, acting as $w_0\delta$
transposed on $H(2)$, the minus vanishing modulo 2) used to implement method
|TitsGroup::dual_twisted| of a |TorusPart|, i.e., element of $H(2)$. The new
method |TitsGroup::dual_twisted| of a |TitsElt| requires an additional
|TorusPart| (shift that is added), which will be computed by the |KGB| class.
Although the |TitsGroup::dual_involution| will only be used for the regular
|TitsGroup| used during |KGB| construction, it is also set to a proper value
for |TitsGroup| instances made by different contructors (used for classifying
real forms, and for handling subsystems

Also add method |GlobalTitsGroup::dual_twisted| of a |TorusElement| (applies
$-w_0$ then $\delta$) or of a |GlobalTitsElement| (which merely takes
|dual_twisted| of |WeylElt| and |TorusElement| components) (while also |using|
the two methods of that name inherited from |TwistedWeylGroup|).

* files kgb.{h,cpp}

The class |KGB_base| keeps just a single |Hermitian_dual| method, implemented
through a |dual| link for each element. However since the map that should be
used when our |KGB| is for the dual real form cannot be derived from the one
used when our |KGB| is for the real form (as used to be assumed, intermixing
cross action by $w_0$) we are obliged to set the |dual| links directly to the
desired result during construction. This requires that constructors for |KGB|
and |global_KGB| be told which case applies; this is achieved by passing them
an additional flag |is_dual| (defaulting to |false| so that code that does not
care about Hermitian duals need not be changed). Implementing the correct dual
twist in |global_KGB| just involves calling both versions of the
|GlobalTitsGroup::dual_twist| method for the |WeylElt| and |TorusElement|
components, but for doing it in |KGB| (the case that is of more interest), an
additional (private) method |KGB::is_dual_twist_stable| tests whether the
required dual twist of the base grading can be represented by a shift in the
|TorusPart| (element of $H(2)$), and if so computes such a shift. The |KGB|
constructor then (after generating the elements) calls this method, and in
case of success tests if calling |TitsGroup::dual_twisted| with a sample
element and the proposed shift produces an element of the generated set; only
then the shift will be used for all dual links, and if not no dual links will
be installed at all. So 3 cases arise: non-dual or dual twist, or none at all.
The method |KGB::is_dual_twist_stable| has the most complicated new code.

* files realredgp.{h,cpp}

Since |Block| has a constructor which receives as argument two objects of type
|RealReductiveGroup|, from which it obtains a |KGB| object, it must be
possible to tell the |RealReductiveGroup| that this |KGB| object should be
constructed as a dual |KGB| for the Hermitian dual links. This is achieved by
complementing the method |RealReductiveGroup::kgb| with another method
|RealReductiveGroup::kgb_as_dual|. Since the |RealReductiveGroup| in fact
stores the |KGB| object it produces, its structure is modified so that it can
separately store two versions (from realex one might use both simultaneously).

* files blocks.{h,cpp}

When constucting a (traditional) |Block|, take care to pass the proper values
of |is_dual| when constructing |KGB| set on the fly, or to properly choose
between methods |RealReductiveGroup::kgb|, |RealReductiveGroup::kgb_as_dual|
when getting them from real reductive groups. The |Block| constructor from two
|KGB| sets is made private to avoid accidental calls with inconsistent twists.
Omit cross acting by $w_0$ when twisting $y$ values for setting the Hermitian
dual links, as the dual |KGB| set now does things properly itself. However,
only set Hermitian dual links at all if |Block| is twist-stable, namely if the
dual |KGB| set has defined Hermitian dual links; also store |orbits| of the
twist on the Dynkin diagram only in this case (facilitating stability test).