Safety commit: adopt usage of |T_coef| method by |KL_table::product_comp|
This should make the twisted KLV computations aware of the signs that have
been inserted into the block. Yet there must be other places where signs are
needed, since 'go' with the corrected block for SO(5,5) triggers an assert
failure (whereas with the uncorrected block with failing braid relations it
does not!). Meanwhile |KL_table::product_comp| could probably now be greatly
simplified, since is uses the same calls to get the coefficients for the
edges, regardless of the descent type that is involved.