Minor changes to interpreter: allow initial underscore, fix output buglet
The lexer is changed to allow identifiers to start with an underscore. The
output for global identifiers that are declared (with a type only) forgot to
indent when from an included file, as all other output did; this is fixed. An
old workaround for a debugger bug was removed from the RootDatum constructor.