Clean up new recursion code for extended KLV; remove bug for defect x case
The formula used involves upward division by $1+q$ with an undertermined top
coefficient, due to the fact that a term has not been incorporated into M[sx]
int his case; however the assert statement that tested the coherence of the
resutl found had a wrong sign. This allows completing the large E6 block.