Remove the DualRealForm type from the realex language; use RealForm instead
The distincition bewteen these two was artificial anyway, and conversions from
no to the other were available to get around type constraints. So in the end
it is better to just rely on runtime checks for compatibility; working with
groups and their dual has become simpler by the chnage.

Also add some functions to basic.rx for consistency of the interface