Move to C++11 for realex, and reorganise that Atlas build process
It was decided to change to C++11 as the programming language for the atlas
project, starting with realex (which is most affected by the change).
The [Mm]akefiles had to be rewitten for this, and the occasion was taken to
clean them up and sanitize the build process. In principle the "make -j4" (or
other number of processes) should now work well for compiling the whole Atlas
software. Much changes were necessary in the realex code, partly because of
changes prototypes for |shared_ptr<T>| constructors.

Also the cwebx inclusion was limitied to a compile-only skeleton, and moved
into the atlas/cwebx directory (initial version of cwebx here is 3.6).