Squat a bug provoked by passge to C+11 that caused realex segmentation fault
This was a really silly bug, caused by the fact that C++11 will no longer
accept static constant integer class members initialised in the class
definition when it has at some point to bind it to a constant reference
argument, forcing a temporary to be created (the linker will fail to find it;
defining the constant again with a value outside the class definition also
gives a problem). The solution that made the compiler shut up, namely define
it as an enumeration value, caused a segmentation fault at startup, because
the (integer) enumeration value did not compare correctly with (unsigned)
|short| values used in the hash table. Finally this has been solved by not
defining the constant value in the class definition, but doing so in a
constant definition outside the class definition.