Refactor code; make parser handle |expr| type by pointer.
The purpose is to make it possible ot transform |expr| into a non POD type.
This involves make complications, splitting many functions in a (new) part
defined in terms of pointers, and an old part in terms of |expr| values. It is
hoped that successive simplifications, and notably the introduction of list
container types will ultimately simplify the code.