Transform |let_list| into a list type, add |singleton| method to sl_list.h
Also make |let| (used in a variant of |expr|) into a smart pointer type.
The |let_list| type is internal to the parser, so changes are limited here,
but the routines for putting together let expressions we restructured. Also
updated the comments (some very old) in the relevant section of parsetree.w
During the change a constructor call of the form |expr(ptr)| occurred where
|ptr| was intended to be a smart pointer, but accidentally was a raw pointer,
which then instead matched the |exp(bool)| constructor after implicit
converion of the poitner to |bool|; this made us decide to delete that
constructor, and to circument using it in the unique place where such use was
actually intended.