Move global identifier table from evaluator,w to global.w, and improve it
Improvements involve using smart pointers inside the table, and creating a
slot with |std::map::emplace|. Also repair a trivial error in main Makefile.
The method |Id_table::add| now uses the |equal_range| idiom to avoid
default-constructing an empty entry in case of a fresh identifier. The
structure |id_data| was made into a class, and the type it holds itself is
stored there, rather than a pointer to it. This forces being more honest about
const-ness: the |Id_table::type_of| method is sometimes used with the
resulting type being subsequently specialised, effectively changing the state
of the table; this usage should be (and now is) marked as manipulator method.
Remove a level of pointer from the final (type) argument of |Id_table::add|.