Declare constants as such even if providing |type| argument to |convert_expr|
The types |void_type|, |int_type|, |bool_type| were declared non-const because
they provided the |type| argument to |convert_expr|. By making local copies as
appropriate, the global variable can now be made |const| so that it is more
evident (no longer dependent on the implementation) that they won't change.
In the same vein the look-up of types of identifiers has been made to return
a pointer to constant type, so that its usage cannot inadvertently modify the
stored type for the identifier; usage as |type| argument to |convert_expr|
now requires making a local copy, and for those cases where specialisation of
the stored type is actually desired, a dedicated method is explicitely called.

Also |Hash_table::id_type| was abbreviated to |id_type| throughout the code