Make bodies of loops in void context have undetermined type rather than void
This should no be restrictive as the loop body is a single expressions. Should
it be a conditional, then having at least one branch have void type suffices
for the others to be forced to be void anyway (and any remaining problems
could be solved using a cast to void). At runtime the void context will
propagate inwards anyway (making the loop body value ignored), and by not
insisting on void in the type checker, useless 'voided:' calls can be avoided.