Complete const-ness transformation of |shared_value|, by including |row_value|
To facilitate the transformation, the pointers held in the vector inside a row
or tuple had been retained as shared pointers to non-const. However, this was
nowhere really necessary, and making them again |shared_value|s (but with the
new definition) was quite easy, and in fact allowed removal of const casts.