Iron out wrinkles in sl_list.h, and write more exhaustive tests in lists.cpp
The |sl_list::ensure| helper class was simplified, limiting it to cases where
it is really helpful. The interface was made conforming by having methods like
|erase| take |const_iterator| values, and return an |iterator|, using a new
explicit conversion in that direction only accessible to the list classes.
Several bugs were caught in sl_list (forgetting to initialise fields, or to
set them in all cases in assignment). Comparing |std::unique_ptr| instances to
|nullptr| systematically calls |get()|, which also helps catching unintended
assignment |=| of the null pointer in place of testing.