Reorganise dual |ComplexReductiveGroup| constructor, to avoid future confusion
The confusion is due to the fact that the |tw| field of |C_info| used to be
changed after the fact, so that attempt to construct the |CartanClass|
directly from the |C_info| constructor would lead to inconsistent data. This
is now corrected by doing the necessary |canoncialize| call on the twisted
involution _before_ passing it to the |C_info| constructor.