Simplify and stabilise computation of |strong_real_form_of|
With stabilise we mean that |real_form(G,involution(x),torus_part(x))| in
realex should return the same realform |G| whenever |x| is a KGB element.
This requires that the initial KGB element should be chosen for "natural" real
forms by the same criteria as they are for synthethic real forms.

Code was repaired to correctly locate an element within the same strong real
form with a desired grading of the (imaginary) simple roots; this cannot be
done (as it used to) by making a shift based just on comparing the current and
desired grading, as the jump might land in a different strong real form.
Instead the |preimage| method now traverses the fiber group, testing for match
of the strong real form and the image under the |fund_f.toAdjoint()| map.

Also a |reduce| method was added to |TorusElement|, because the reflections
used when applying (complex) cross actions left the entries at nonstandard
congruence class representatives, giving trivially different cocharacters.