Ensure |strong_real_form_of| returns a unique strong involution representative
Previously it was already ensured that for torus factors of KGB elements in
each real form, then |strong_real_form_of| returns the same strong involution.
However nuisance was still caused by the liberty to represent this strong
involution in different ways. Currently the cocharacter is made to be
represented by a delta-fixed coweight, then reduced modulo 2Z^n (standard
representative of a torus element), and finally minimised (for some criterion)
in its (1+^delta).H(2) coset. It is hoped that this produces a result
independent of the original representative of the strong involution.