Modify testrun (easier reading). Add |central_fiber| and |initial_torus_bits|.
The |CoveringIterator::kernel_generators| method, intended to be used to find
atlas input that will lead to a given root datum, is incompletely implemented.
Reversing the kernel-generator-to-sublattice is subtle and I need to think
more about it more. A |PreRootData| constructor was modified to factor out the
quotient operation (replacing weight lattice by a sublattice) as something to
be explicitly applied after the constructor, for better readablility.