Change choice of torus bits for initial element in full KGB construction
The choice used to be made rather obscurely in the |TitsCoset::grading_seed|
method called from a |EnrichedTitsGroup| derived object. The choice, when it
is at the fundamental fiber, is now made more directly by calling the method
|ComplexReductiveGroup::x0_torus_part|. In this case it is made using the same
criterion of as is made in |strong_real_form_of| to select one cocharacter
value (base grading vector) among the candidates for a given $\xi^2$ value.
Ultimately this means the initial torus bits of a (weak) real form reflect the
difference between the selected cocharacter in the fundamental fiber of the
real form and the purely grading-based cocharacter for its square class.

The result is that those initial torus bits may by themselves look rather
arbitrary; for instance it might happen that there is an element in the
fundamental fiber with all torus bits 0 that is not chosen to become the
initial element of the KGB construction. However the resulting choices should
make synthethic and selected real forms more regular, and behave more alike.