Add user type abbreviations to realex
One can now define type abbreviations in realex using the syntax
: type_name = type_expression

Here type_name should be a fresh identifier, not yet used as global variable
or as (overloaded) function name. It will henceforth abbreviate the type
expression. For now this is an input-only affair, any feedback from realex
will continue to use the expanded version of the type.

At the same time some general clean-up of global.w was done (serveral methods
had interfaces that were inconsistent and/or could be improved).