Modify some internal structures to facilitate access to input files names
The recording of file names in |BufferedInput| is now a two-stage affair: the
file names are entered into the hash table when the file is sucessfully
opened, and they are marked as completed (so they won't normally be opened
again) once they have been sucessfully read. As a consequence the file names
recorded in the |input_record| structure can now be stored as numbers.