Modify |BufferedInput| so file record destructor deletes its |ifstream| object
This was made possible now that the stack is implemented as a list (in fact a
|mirrored_sl_list|) rather than as a |std::vector|, which allows the stack
records to forbid (and therefore not implement) any form of copying. This does
require that instead of |push_back| one uses |emplace_back| (a method that had
to be added to |mirrored_sl_list|) to create a new record on the stack.