Some changes to basic.rx and consequences; also update Vogan-dual.rx
The modifications to Vogan-dual.rx involve adding fairly complicated maneuvers
to ensure in computing a dual KGB element that a theta-stable (on the dual
side) candidate for the torus factor in the required X_*-coset (that is, with
the correct value for its square) is computed. Whether it is really valid
seems to depend on finding a theta-stable solution for finding an integral
solution to the theta-stable problem of giving the right gradings, so it is
possible that the maneuvers aren't sufficiently complicated yet.