Stratify the parse tree construction and eliminate all |mk_...| functions
Reorganise formula stack to attach source location to all operator symbols
Provide arguments to functions so as to make a location structure available
wherever |expr| values are constructed, and add a |source_location| structure
to the non-variant part of |expr|, which is in all cases initialised from that
argument. For now changes are insufficent on getting the *correct* locations
everywhere, notably in formulae, and empty argument lists have no location.