Change "precedence" of ';' to equal that of 'next', allowing 'let' to follow ';'
It turns out it is much more practical to have the precedence of ';' to be at
the lowest level, that of 'let' clauses, functions, casts and so forth; this
avoids having to open a group after ';' if it is a let-expression or function.

Having both ';' and 'next' have the same (lowest) binding power while
associating to the right seems quite natural: in a sequence of expressions
linked by these separators only, the value will be given by the expression
preceeding the first 'next', if any, and otherwise the by the final expression.