Fix what seems like a transcription error and some true errors in synthetic.rx
Two cases involve using id where rd is required (as is corrected in Jeff's
file synthetic_aux.rx). In the diff vectors this must be a transcription
error, but using id in make_simple suggest that one was not trying to only
move to an integrally simple situation; however for Cayley one does need to
use a purely simple root.

A third correction is made to the "projection" part in posroot_inv_Cayley,
which seemed to use a wrong formula: gamma-proj should be lambda(p)+proj