Make handling of parameters more robust with respect to dominance questions.
Notably when subscripting a ParamPol value by a Param, make sure the latter is
transformed to the precise form in which it would appear in a ParamPol, and
signal an error if it is zero or not final (non standard parameters will also
cause an error). For equality testing of parameters, don't simply refuse (as
was the case) to consider non standard parameters, but test them for strict
equality, reserving equivalence for standard parameters. Finally optimise the
operation of mutiplying a ParamPol to the left by an integer 0.