Change syntax of anonymous parameterless functions to use '@:...'; fix hecke.rx
The old syntax was '():...', but that would cause conflict when ':' gets used
(soon) in array slices, and could follow '()'. The replacement '@' for '()'
seems the least ugly alternative, and does not cause new conflicts in parser.

The script polynomial.rx now restores ordinary vector arithmetic at the end of
its loading, so that laoding it does not perturb other scripts; the implicit
zero-stripping now done in polynomial subtraction did break some scripts.

Consequently hecke.rx needed fixing, as it depended on the implicit switch to
polynomial arithmetic. But is also uses ordinary vector (and ratvec)
arithmetic, so it was necessary to carefully separate vector and polynomial
arithmetic operations (all polynomial arithmetic happens inside |hecke_row|).