Implement reversed indexing for row, vec, ratvec, string, mat (entry or column)
Mixed matrix index/slice will not be in realex 0.9: full column or row
(through transposition) selection is possible; for a partial slice just select
then trim the selection. The only thing left to do in this domain is to allow
slicing for matrices, provided a pair of (int,int) lower/upper bounds are used.