Allow marking identifiers as constant, and omitting unused loop variables
Constants identifiers are marked by preceding them by '!'. Currently the
constness of global variable is not implemented, although it is they that
could actually benefit somewhat from constness (avoiding the shared_share).
Local identifiers can be modifiable at no cost it turns out, so currently
there is no point in making them const; for interation variables it even
happened that nilpotent.rx had a valid use for modifying one, which had to be
rewritten to comply with the new constness requirement (ah, bureaucracy).

This case also highlighted that reverse-indexing should be allowed in
component updates just as it is in selections; more things to do.